+++ RELEASE-NOTES.md
@@ -31,5 +31,7 @@
 - Mobile styling updates
 
 ## Unreleased
+- reload page on account change
 - Address page updates
 - Wallet modal fix very small balances
+