+++ RELEASE-NOTES.md
@@ -34,4 +34,4 @@
 - reload page on account change
 - Address page updates
 - Wallet modal fix very small balances
-
+- change "cancel" to "close" in wallet modal