+++ RELEASE-NOTES.md
@@ -31,4 +31,5 @@
 - Mobile styling updates
 
 ## Unreleased
-Nothing at this time
+- displayed values for token0/token1 in staked balance accuracy improved
+- ubq Price oracle timeframe expanded to reduce potential for missing the latest price